reorganized subpackages and removed scratch script
