+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023/preconversion/extract_metadata.py
